Исправления
Диапазон опыта расширен.
Корабли не берут бонусы с других полей.
Все ресурсы заменены на дерево.

+++ Assets/Resources/Scripts/ScreenLevel.cs
@@ -12,7 +12,7 @@ public class ScreenLevel : MonoBehaviour
     public int minLevel, maxLevel;
 
     private Island island;
-    private int oldExp;
+    private long oldExp;
     private bool filled = false;
 
     private void Awake()
